Added an instructions UI and scene loaders
So I added instructions to the button presses. I separated the scenes so there would be multiple scenes, one for the puzzle section where the player has to find the key to unlock the drawer to find the combination, then I am making a scene so that the player can choose which game they'd like to play, the wallsmasher one or the bowling game. I also need to add a back button to my game and you should add one to yours as well. As well as a restart button in the gameselection screen if they want to play the puzzle one again.